added new columns and fixed a bug
I added cue, content, and source counts and in doing so I realized the regex I used for identifying also captured some edge cases it shouldn't have. So that's been fixed now and apparently that somehow made the code close to ten times as fast as it was before. I really don't want to look into why that might be, but the numbers still seem fine